requirements_torch.txt
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
Pillow==8.1.0
plyfile==0.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
torch==1.7.1
trimesh==3.9.1
typing-extensions==3.7.4.3
neptune-client==0.5.0
scikit-learn==0.24.1
tqdm==4.57.0
PyMCubes==0.1.2
Cython==0.29.22
pyglet==1.5.15